# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Carteira Financeira",
        "description": "API RESTful para gerenciamento de carteiras, depósitos e transferências.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor API Local"
        }
    ],
    "paths": {
...
        "/api/deposit": {
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Realizar um depósito na carteira do usuário",
                "operationId": "7d37929cec1ed6739970c3ba17c8102b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Depósito de salário",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Depósito realizado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Depósito realizado com sucesso!"
                                        },
                                        "new_balance": {
                                            "type": "string",
                                            "format": "float",
                                            "example": "223.95"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Ação proibida (carteira com inconsistência)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transfer": {
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Realizar uma transferência para outro usuário",
                "operationId": "ff447720cf8647ed4d7a3854209f0b69",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "to_user_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "destino@example.com"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Pagamento de conta",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transferência realizada com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transferência realizada com sucesso!"
                                        },
                                        "your_new_balance": {
                                            "type": "string",
                                            "format": "float",
                                            "example": "173.95"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Ação proibida (carteira de destino com inconsistência)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transações"
                ],
                "summary": "Listar histórico de transações do usuário autenticado",
                "operationId": "6581469356772e7a74a0d79897544f1a",
                "responses": {
                    "200": {
                        "description": "Histórico de transações recuperado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/{id}/reverse": {
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Reverter uma transação (depósito ou transferência)",
                "operationId": "d7609df695f176ccc7a9a83524c299c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da transação a ser revertida",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Erro de digitação no valor.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transação revertida com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transação #1 revertida com sucesso!"
                                        },
                                        "original_transaction_status": {
                                            "type": "string",
                                            "example": "reversed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida (transação já revertida, falhou ou tipo não suportado)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor (ex: saldo insuficiente para reversão)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthorizedError": {
                "title": "Erro de Autenticação",
                "description": "Resposta para requisições não autenticadas.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "ForbiddenError": {
                "title": "Acesso Proibido",
                "description": "Resposta para ações proibidas (ex: carteira com inconsistência).",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Não é possível depositar. Sua carteira está com uma inconsistência que impede novos depósitos. Por favor, contate o suporte."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Erro de Validação",
                "description": "Resposta para falhas de validação de dados.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field_name": [
                                "The field_name field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "InternalServerError": {
                "title": "Erro Interno do Servidor",
                "description": "Resposta para erros inesperados do servidor.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Ocorreu um erro interno do servidor."
                    },
                    "error": {
                        "type": "string",
                        "example": "Detalhes do erro para debug (remover em produção)"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}